backend/requirements.txt
fastapi
uvicorn
langchain
langchain-mistralai
python-dotenv
pydantic
langchain-core
langchain-community
